inspektor: Rename check to patch
